+++ allennlp_models/pair_classification/dataset_readers/snli.py
@@ -2,7 +2,6 @@ from typing import Dict, Optional
 import json
 import logging
 
-from overrides import overrides
 
 from allennlp.common.file_utils import cached_path
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
@@ -73,7 +72,6 @@ class SnliReader(DatasetReader):
             self._combine_input_fields = isinstance(self._tokenizer, PretrainedTransformerTokenizer)
         self.collapse_labels = collapse_labels
 
-    @overrides
     def _read(self, file_path: str):
         # if `file_path` is a URL, redirect to the cache
         file_path = cached_path(file_path)
@@ -88,7 +86,6 @@ class SnliReader(DatasetReader):
                 hypothesis = example["sentence2"]
                 yield self.text_to_instance(premise, hypothesis, label)
 
-    @overrides
     def text_to_instance(
         self,  # type: ignore
         premise: str,
@@ -121,7 +118,6 @@ class SnliReader(DatasetReader):
 
         return Instance(fields)
 
-    @overrides
     def apply_token_indexers(self, instance: Instance) -> Instance:
         if "tokens" in instance.fields:
             instance.fields["tokens"]._token_indexers = self._token_indexers