+++ allennlp_models/rc/dataset_readers/squad.py
@@ -1,8 +1,6 @@
 import json
 import logging
-from typing import Any, Dict, List, Tuple, Optional
-
-from overrides import overrides
+from typing import Any, Dict, List, Tuple, Optional, Iterable
 
 from allennlp.common.file_utils import cached_path
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
@@ -100,7 +98,6 @@ class SquadReader(DatasetReader):
         self.skip_impossible_questions = skip_impossible_questions
         self.no_answer_token = no_answer_token
 
-    @overrides
     def _read(self, file_path: str):
         # if `file_path` is a URL, redirect to the cache
         file_path = cached_path(file_path)
@@ -143,13 +140,12 @@ class SquadReader(DatasetReader):
                     if instance is not None:
                         yield instance
 
-    @overrides
     def text_to_instance(
         self,  # type: ignore
         question_text: str,
         passage_text: str,
         is_impossible: bool = None,
-        char_spans: List[Tuple[int, int]] = None,
+        char_spans: Iterable[Tuple[int, int]] = None,
         answer_texts: List[str] = None,
         passage_tokens: List[Token] = None,
         additional_metadata: Dict[str, Any] = None,