+++ allennlp_models/vision/models/heads/nlvr2_head.py
@@ -1,7 +1,7 @@
 from typing import Dict, Optional
 
 import torch
-from overrides import overrides
+
 
 from allennlp.data.vocabulary import Vocabulary
 from allennlp.models.heads.head import Head
@@ -25,7 +25,6 @@ class Nlvr2Head(Head):
         self.accuracy = CategoricalAccuracy()
         self.fbeta = FBetaMeasure(beta=1.0, average="macro")
 
-    @overrides
     def forward(
         self,  # type: ignore
         encoded_boxes: torch.Tensor,
@@ -55,7 +54,6 @@ class Nlvr2Head(Head):
 
         return output
 
-    @overrides
     def get_metrics(self, reset: bool = False) -> Dict[str, float]:
         result = self.fbeta.get_metric(reset)
         result["accuracy"] = self.accuracy.get_metric(reset)